Move util method to ArgumentsTests class
